heatline-pi-fastapi/requirements.txt
fastapi==0.115.12
uvicorn[standard]==0.34.0
pydantic==2.11.3
requests==2.32.3
python-multipart==0.0.20
numpy==2.2.4
Pillow==11.1.0
opencv-python-headless==4.11.0.86
